fix(solve): 존 배정 DP 호환 마스크 int → long (차종 ≤ 64)
기권을 없앤 직전 커밋이 종전 기권선(Π > 65,536 → 차종 ≥ 17에서 기권)이 사실상
맡고 있던 마스크 폭 가드를 치웠다. `Demand`의 호환 마스크가 `int`라 유형 t의 비트
`1 << t`가 t ≥ 32에서 t − 32로 접히고, **예외 없이 답만 틀어진다** — 33종 입력에서
호환되지 않는 차가 존에 배정된다.

- `Demand.compat`을 `long`, `prefixMask`·`groupMask`를 `long[]`로 넓히고
  모든 `1 << t`를 `1L << t`로, 그룹 map을 `TreeMap<Long, long[]>`·`Long::bitCount`로 바꿨다.
- **불변식: 차종 수 ≤ 64.** 65종 이상은 같은 aliasing이 재발하는 알려진 한계로
  X28에 등재만 한다 — 가드·기권·예외를 두지 않는다 (H23·H24의 `abstains == false`는 계약).
- T57 신설 — 차종 33개, ZBIG은 유형 32만 허용하는 입력. 마스크가 접히면 낭비가 작은
  유형 0의 차가 ZBIG에 들어와 깨진다(실제로 `1L << (t % 32)`로 접어 보면 V000이 배정된다).
  `1L << t` 치환 누락은 경고가 없으므로 이 시험이 유일한 방어선이다. T53~T56은 H25 대역이라 비웠다.

부작용은 **차종이 정확히 32개**인 입력뿐이다 — `int`에서 `1 << 31`이 음수라 31번 유형이 낀
마스크가 그룹 정렬에서 맨 앞으로 갔는데 `long`에서는 뒤로 간다. 동률 갈림이 달라질 수 있고
둘 다 유효하다(오답이 아니다). 32종을 쓰는 기존 시험은 없다. 31종 이하와 실물 6종은 순서 불변이고,
실물 fixture score `[0, 31, 4,198,408, 1,002,069]`·`truncated == false`는 그대로다 (T52).

+++ solver-core/src/main/java/com/ronext/rpdptw/solve/ZoneQuotaAllocation.java
@@ -213,7 +213,7 @@ final class ZoneQuotaAllocation {
     /** 존이 유형 t를 최대로 쓸 수 있는 대수 — 호환 아니면 0, 아니면 자원별 올림의 최대 (최소 1). 용량·한도 부재는 그 항 0. */
     static int maxNeed(Demand demand, VehicleType type) {
         int index = demand.types.indexOf(type);
-        if (index < 0 || (demand.compat & (1 << index)) == 0) {
+        if (index < 0 || (demand.compat & (1L << index)) == 0) {
             return 0;
         }
         long need = 1L;
@@ -532,18 +532,18 @@ final class ZoneQuotaAllocation {
         final long totalVolume;
         final long totalWeight;
         final long totalVisits;
-        final int compat;
-        private final int[] prefixMask;
+        final long compat;                                                  // 유형 t의 비트 = 1L << t (차종 ≤ 64)
+        private final long[] prefixMask;
         private final long[] prefixVolume;
         private final long[] prefixWeight;
         private final long[] prefixVisits;
-        private final int[] groupMask;
+        private final long[] groupMask;
         private final long[] groupMaxVolume;
         private final long[] groupMaxWeight;
 
-        private Demand(List<VehicleType> types, long totalVolume, long totalWeight, long totalVisits, int compat,
-                       int[] prefixMask, long[] prefixVolume, long[] prefixWeight, long[] prefixVisits,
-                       int[] groupMask, long[] groupMaxVolume, long[] groupMaxWeight) {
+        private Demand(List<VehicleType> types, long totalVolume, long totalWeight, long totalVisits, long compat,
+                       long[] prefixMask, long[] prefixVolume, long[] prefixWeight, long[] prefixVisits,
+                       long[] groupMask, long[] groupMaxVolume, long[] groupMaxWeight) {
             this.types = types;
             this.totalVolume = totalVolume;
             this.totalWeight = totalWeight;
@@ -559,15 +559,15 @@ final class ZoneQuotaAllocation {
         }
 
         static Demand of(Problem problem, List<VehicleType> types, Zone zone) {
-            TreeMap<Integer, long[]> groups = new TreeMap<>(
-                    Comparator.comparingInt(Integer::bitCount).thenComparing(Comparator.naturalOrder()));
+            TreeMap<Long, long[]> groups = new TreeMap<>(
+                    Comparator.comparingInt(Long::bitCount).thenComparing(Comparator.naturalOrder()));
             long total = 0L;
             for (RequestId id : zone.members()) {
                 Request request = problem.request(id);
-                int mask = 0;
+                long mask = 0L;
                 for (int t = 0; t < types.size(); t++) {
                     if (types.get(t).compatible().contains(id)) {
-                        mask |= 1 << t;
+                        mask |= 1L << t;
                     }
                 }
                 long[] g = groups.computeIfAbsent(mask, m -> new long[5]);       // Σvolume, Σweight, Σvisits, max volume, max weight
@@ -579,19 +579,19 @@ final class ZoneQuotaAllocation {
                 total = Math.addExact(total, request.totalVolume());
             }
             int n = groups.size();
-            int[] prefixMask = new int[n];
+            long[] prefixMask = new long[n];
             long[] prefixVolume = new long[n];
             long[] prefixWeight = new long[n];
             long[] prefixVisits = new long[n];
-            int[] groupMask = new int[n];
+            long[] groupMask = new long[n];
             long[] groupMaxVolume = new long[n];
             long[] groupMaxWeight = new long[n];
             int k = 0;
-            int mask = 0;
+            long mask = 0L;
             long volume = 0L;
             long weight = 0L;
             long visits = 0L;
-            for (Map.Entry<Integer, long[]> entry : groups.entrySet()) {
+            for (Map.Entry<Long, long[]> entry : groups.entrySet()) {
                 mask |= entry.getKey();
                 volume = Math.addExact(volume, entry.getValue()[0]);
                 weight = Math.addExact(weight, entry.getValue()[1]);
@@ -616,7 +616,7 @@ final class ZoneQuotaAllocation {
                 long visits = 0L;
                 boolean unlimitedVisits = false;
                 for (int t = 0; t < types.size(); t++) {
-                    if (s[t] == 0 || (prefixMask[k] & (1 << t)) == 0) {
+                    if (s[t] == 0 || (prefixMask[k] & (1L << t)) == 0) {
                         continue;
                     }
                     VehicleType type = types.get(t);
@@ -633,7 +633,7 @@ final class ZoneQuotaAllocation {
                 }
                 boolean fitsLargest = false;
                 for (int t = 0; t < types.size() && !fitsLargest; t++) {
-                    fitsLargest = s[t] > 0 && (groupMask[k] & (1 << t)) != 0
+                    fitsLargest = s[t] > 0 && (groupMask[k] & (1L << t)) != 0
                             && types.get(t).maxVolume() >= groupMaxVolume[k] && types.get(t).maxWeight() >= groupMaxWeight[k];
                 }
                 if (!fitsLargest) {
